whisper-server/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
flask>=2.3.0
flask-cors>=4.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
python-multipart>=0.0.6
gunicorn>=21.0.0
python-dotenv>=1.0.0